+++ src/sections/perps/poweredBy.tsx
